# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para sugestões de presentes personalizadas.",
        "title": "Gift App API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/users": {
            "put": {
                "description": "Exemplo de payload: {\"fullName\":\"Ana Souza\",\"email\":\"ana.souza@email.com\",\"active\":true,\"planId\":\"basic\",\"birthDate\":\"1992-07-21\",\"city\":\"Sao Paulo\"}. Campo birthDate no formato YYYY-MM-DD.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Criar usuário",
                "parameters": [
                    {
                        "description": "Dados do usuário",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UserUpsertRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/email": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Buscar usuário por email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email do usuário",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{userId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Buscar usuário",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do usuário",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.User"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Exemplo de payload: {\"fullName\":\"Ana Souza\",\"email\":\"ana.novo@email.com\",\"active\":true,\"planId\":\"pro\",\"birthDate\":\"1992-07-21\",\"city\":\"Campinas\"}. Campo birthDate no formato YYYY-MM-DD.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Atualizar usuário",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do usuário",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados a atualizar",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UserUpsertRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.User": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "birthDate": {
                    "type": "string",
                    "format": "date-time",
                    "example": "1999-03-16T00:00:00Z"
                },
                "city": {
                    "type": "string",
                    "example": "Santo Andre"
                },
                "email": {
                    "type": "string",
                    "example": "guilherme.jer1@gmail.com"
                },
                "fullName": {
                    "type": "string",
                    "example": "Guilherme Jeronymo"
                },
                "planID": {
                    "type": "string",
                    "example": ""
                },
                "userID": {
                    "type": "string",
                    "example": "5581a365-394f-467d-ae13-0d01e4cf1863"
                }
            }
        },
...
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "birthDate must be in format YYYY-MM-DD"
                }
            }
        },
...
        "http.UserUpsertRequest": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "birthDate": {
                    "type": "string",
                    "format": "date",
                    "example": "1992-07-21"
                },
                "city": {
                    "type": "string",
                    "example": "Sao Paulo"
                },
                "email": {
                    "type": "string",
                    "example": "ana.souza@email.com"
                },
                "fullName": {
                    "type": "string",
                    "example": "Ana Souza"
                },
                "planId": {
                    "type": "string",
                    "example": "basic"
                }
            }
        }
    }
}